Heater, Window and AC Cards integrated. Controller created.
Three cards have been made for displaying current statuses of the Heater, the Window and the AC.

A new TemperatureController has been made which is going to handle passing the data from the Temperature Model to the view.

It already has some rudimentary logic to display a warning if some energy inefficient combinations are currently true, such as the heater being on with the window open.