Heater Historical Data now shows in an improved format.
We are now displaying when the heater was turned on and how long it was turned on for. This accoutns for multiple consecutive "on" statuses in a row within the database and will look for the next corresponding "off" status to calculate the amount of time the heater spent on.