Streamlining button to show Historical Data
Showing historical data is now a matter of clicking a single togglable button, which will show whether it is active or not via a blue colour. This is far more intuitive than it was before.

We are also preventing historical records from being shown if they are less than 60 seconds long as this is not really required.