Improved Temperature Chart Timings
Before, the chart would update every time the pusher event was fired. This led to lots of erroneous data points whenever a different client loaded the page as they would also hit the pusher event that requested data.

Now we are simply storing the new data on the client when it is received from pusher, and only updating the chart every 10 seconds, keeping a regular interval between datapoints.