Fixed Historical Data not Showing
Because we were only taking 100 records, it seemed to be ignoring anything with an ID higher than 100.

This solution isn't great. As the table expands, it will have to process every single record, which could result in thousands or more operations per table, but I don't have enough time to make a cleaner solution, so this will have to do for now.