Huge Optimisation
An enormous optimisation has been made on the node-red side, we are now only storing records in the status tables where there has been a change. This could reduce the number of records to record how long a utility has been on for from an infinite amount down to 2.

In line with this, we are now checking only 100 records, which means the maximum number of historical records which will be shown is likely to be roughly 50, potentially a bit less.